ISEC-94: Default and async Routemaster publishers (#56)
* ISEC-94: Default and async Routemaster publishers

While integrating the publishing library I found that:

* It was missing async publishing (and in my case linking availability
of the service to Routemaster by putting it on a hot path, which is
bad).

* I was repeating a bunch of boilerplate from the docs which could be
made into code that doesn’t need to be repeated.

As such, this adds async publishing as an option. This is at the
publisher level rather than in the model, because it’s possible you
might want multiple publishers for a model and one might be sync with
the other async.

It also adds the ability to register default publishers because most of
the time they’ll be the same and you can infer the URL in a common
publisher. I’ve changed the name to `ApplicationPublisher` in the docs
rather than `BasePublisher` to fit in with the general Rails naming
scheme.

* ISEC-94: Add a CHANGELOG entry for async publishing

* ISEC-94: Minor documentation fix for require statement