Add docs for Hirefire for Web Dynos (#70)
* Add docs for Hirefire for Web Dynos

I've added a note in the README for adding HireFire integration for Web dynos (as well as sidekiq users). This _could_ be turned into a RooOnRails plugin, but I figured I'd start with docs. (See SETI-795)

* Additional information

Additional information about setting up the log drain (just in case RooOnRails hasn't done it for you)