Add ability to create exceptions to SSL enforcement
As we move more services onto AWS, we need to serve a health check
over HTTP that returns an 200 status code. At the moment, this
doesn't work for apps using `roo_on_rails` – `Rack::SslEnforcer`
redirects the health check request, which means it returns a 301
status code, and the health check fails.

This adds an environment variable – the ever-so-catchily named
`ROO_ON_RAILS_SSL_ENFORCEMENT_EXCEPTIONS` – that allows developers
to specify a comma-delimited list of exact paths that should be
excluded from SSL enforcement.