PACC-495: Use internal URLs for keys (#106)
While decoding keys within the Deliveroo ecosystem we can use an internal URL instead of an external one (and retrieve the data much faster, and not be subject to the firewall). The Populate Env From JWT library now accepts a list of mapped URLs, with a sensible default that will work with staging and production.