[EXPERIMENTAL] Environment Validation
This adds the ability to specify an Envfile. This file contains a list of required environment variables, with a little logic for validating them/providing defaults. It also can ensure that certain variables are set in Hopper, rather than a dotenv file. In the event that a variable isn't set it will fail hard, and cause a Hopper rollback.